+++ amm-pool-contract/tests/calibrate_gap.rs
@@ -1,13 +1,14 @@
 #![cfg(not(feature = "sdk20"))]
 
+mod common;
+
 #[cfg(test)]
 mod calibrate_gap {
     use amm_pool_contract::ConstantProductPoolClient;
     use soroban_sdk::Env;
 
     fn measure_do_expensive_work(env: &Env) {
-        let wasm_path = "../target/wasm32v1-none/release/amm_pool_contract.wasm";
-        let wasm = std::fs::read(wasm_path).expect("WASM file not found, did you run cargo build?");
+        let wasm = crate::common::load_contract_wasm("wasm32v1-none");
         #[allow(deprecated)]
         let contract_id = env.register_contract_wasm(None, wasm.as_slice());
         let client = ConstantProductPoolClient::new(env, &contract_id);